Completed input menu

+++ SegreteriaStudenti/src/Esami/Esame.java
@@ -6,8 +6,8 @@ public class Esame {
 	private String nome;
 	private int crediti;
 	private int voto = 0;;
-	private static double idUltimo = 0;
-	private double id;
+	private static int idUltimo = 0;
+	private int id;
 	boolean superato = false;
 	
 	//costruttore
@@ -21,7 +21,7 @@ public class Esame {
 	//getter e setter
     public String getNome() { return this.nome; }
     public int getCrediti() { return this.crediti; }
-    public double getId() { return this.id; }
+    public int getId() { return this.id; }
     public int getVoto() {return this.voto; }
     public void setCrediti(int crediti) { this.crediti = crediti; }
     public void setVoto(int voto) {this.voto = voto; }
